+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Fixed
+
+-   Error code propagation through Nordic's Toolchain Manager.
+
 ## [0.3.0] - 2022-10-05
 
 ### Added